habitus-backend-python/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn==0.23.2

# Database
motor==3.3.1
pymongo==4.5.0

# Authentication
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
cryptography==41.0.4
bcrypt==4.0.1

# Data processing
pandas==2.1.1
openpyxl==3.1.2
xlrd==2.0.1
numpy==1.26.0

# Environment and settings
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0

# Email
aiosmtplib==2.0.2
jinja2==3.1.2

# Utilities
python-dateutil==2.8.2
email-validator==2.0.0
ujson==5.8.0

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.25.0

# Development
black==23.9.1
isort==5.12.0
mypy==1.5.1
flake8==6.1.0
